codonopt/requirements.txt
biopython
pandas
openpyxl
tqdm
numpy
cryptkeeper